+++ package.json
@@ -22,19 +22,19 @@
   "dependencies": {
     "@sindresorhus/slugify": "^2.1.0",
     "debounce": "^1.2.1",
-    "gatsby": "^4.7.1",
-    "gatsby-plugin-gatsby-cloud": "^4.7.0",
-    "gatsby-plugin-google-analytics": "^4.7.0",
-    "gatsby-plugin-image": "^2.7.0",
-    "gatsby-plugin-react-helmet": "^5.7.0",
-    "gatsby-plugin-sharp": "^4.7.0",
-    "gatsby-plugin-sitemap": "^5.7.0",
+    "gatsby": "^4.9.0",
+    "gatsby-plugin-gatsby-cloud": "^4.9.0",
+    "gatsby-plugin-google-analytics": "^4.9.0",
+    "gatsby-plugin-image": "^2.9.0",
+    "gatsby-plugin-react-helmet": "^5.9.0",
+    "gatsby-plugin-sharp": "^4.9.0",
+    "gatsby-plugin-sitemap": "^5.9.0",
     "gatsby-plugin-web-font-loader": "^1.0.4",
-    "gatsby-source-contentful": "^7.5.0",
-    "gatsby-source-filesystem": "^4.7.0",
-    "gatsby-source-shopify": "^6.7.0",
-    "gatsby-transformer-remark": "^5.7.0",
-    "gatsby-transformer-sharp": "^4.7.0",
+    "gatsby-source-contentful": "^7.7.0",
+    "gatsby-source-filesystem": "^4.9.0",
+    "gatsby-source-shopify": "^6.9.0",
+    "gatsby-transformer-remark": "^5.9.0",
+    "gatsby-transformer-sharp": "^4.9.0",
     "isomorphic-fetch": "^3.0.0",
     "lodash.debounce": "^4.0.8",
     "lodash.isequal": "^4.5.0",
@@ -45,6 +45,6 @@
     "react-icons": "^4.3.1",
     "sanitize-html": "^2.7.0",
     "shopify-buy": "^2.14.0",
-    "urql": "^2.1.3"
+    "urql": "^2.2.0"
   }
 }
