+++ public/page-data/404.html/page-data.json
@@ -2,4 +2,4 @@
     "componentChunkName": "component---src-pages-404-js",
     "path": "/404.html",
     "result": {"pageContext":{}},
-    "staticQueryHashes": ["1096730463","2252784942","3277454263"]}
+    "staticQueryHashes": ["2252784942","3234233177","3277454263"]}
