+++ src/pages/misc/{ContentfulEditorial.id}.js
@@ -15,14 +15,14 @@ export default function Page({data}) {
     localStorage.setItem("lang", "fr");
   }
 
-  const node = data.contentfulPageSimple;
+  const node = data.contentfulEditorial;
 
   if (node.contenu !== undefined) {
     return (
       <Layout>
         <div className={container}>
           <div className={article}>
-            <h1>{node.titre}</h1>
+            <h1>{node.sousTitre}</h1>
           </div>
           
           <div className={container_image}>
@@ -40,8 +40,10 @@ export default function Page({data}) {
 
 export const query = graphql`
   query($id: String!) {
-    contentfulPageSimple(id: {eq: $id}) {
+    contentfulEditorial(id: {eq: $id}) {
+      id
       titre
+      sousTitre
       contenu {
         childMarkdownRemark {
           html
@@ -57,8 +59,6 @@ export const query = graphql`
     }
   }
 `;
-
-
 /*
 export default function Page({data}) {
   const brownser_is = typeof window !== "undefined";
